Saves the ID and Question of a decision before creating it
The first two fields in a row that our parser is traversing through are
the ID and Question field. Therefore we need to save these two before
getting to the decisionType field and creating the correct decision. We
are then able to give these two values already in the constructor.
We could also just save everything first as a variable and then create
everything on exitRow, but currently the first approach is being
implemented.