Created whole new parser structure and added licenses
Changed combined grammar to seperate lexer and parser grammar to ensure
a better structure and make antlr mode use possible. Even though it is
not needed for the json file parsing.

The new seperated grammar includes a finished and proper parsing of json
and csv dopler models depending on which file is being chosen.
The lexer grammar checks if it can find as first character an '{' to see
if its a json file. Otherwise it assumes the file is a CSV file.