Implemented entirely new json and csv writer and some bugfixes
the first writer made many extra symbols that were inconsistent and not
compatible with the antlr grammar. This has been fixed by using standard
FileWriter and writing everything manually in the intended syntax.

Fixed some bugs that came up when trying to read a model that was
created with the writer.